test: 모든 게시글 조회 findArticles 테스트 코드 추가

+++ src/main/java/com/example/onemoreboard/service/ArticleService.java
@@ -6,8 +6,10 @@ import com.example.onemoreboard.entity.Article;
 import com.example.onemoreboard.repository.ArticleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * packageName    : com.example.onemoreboard.service
@@ -42,8 +44,8 @@ public class ArticleService {
 
 
     public ArticleResponse updateShowArticle(Long id) {
-      Article article = articleRepository.findById(id).orElse(null);
-      return ArticleResponse.fromEntity(article);
+        Article article = articleRepository.findById(id).orElse(null);
+        return ArticleResponse.fromEntity(article);
     }
 
 
@@ -57,7 +59,41 @@ public class ArticleService {
     }
 
     public void deleteArticleById(Long id) {
-        articleRepository.deleteById(id);
+        Article article = articleRepository.findById(id).orElse(null);
+
+        if (article != null) {
+            articleRepository.deleteById(id);
         }
+        throw new RuntimeException("게시글이 없습니다.");
     }
+
+    // 타임리프 업데이트방식이랑 RestApi 업데이트방식이 다르다.
+    public ArticleResponse restUpdateArticle(Long id, ArticleRequest articleRequest) {
+        Article article = articleRepository.findById(id).orElse(null);
+
+        if (article != null) {
+           article.updateArticle(articleRequest.getTitle(),articleRequest.getContent());
+           articleRepository.save(article);
+           return ArticleResponse.fromEntity(article);
+        }
+        return null;
+    }
+
+    // 트랜잭션 테스트
+    @Transactional
+    public List<Article> test(List<ArticleRequest> articleRequestList) throws IllegalAccessException {
+        List<Article> article = articleRequestList.stream()
+                .map(dto -> dto.toEntity())
+                .collect(Collectors.toList());
+
+        article.stream()
+                .forEach(articles -> articleRepository.save(articles));
+
+        articleRepository.findById(-1L)
+                .orElseThrow(()-> new RuntimeException("결제실패"));
+
+        return article;
+    }
+}
+
